Moved SignOutWordPressComAsync from WordPress to MeFragment
