+++ WordPress/src/main/java/org/wordpress/android/WordPress.java
@@ -3,7 +3,6 @@ package org.wordpress.android;
 import android.annotation.TargetApi;
 import android.app.Activity;
 import android.app.Application;
-import android.app.ProgressDialog;
 import android.content.ComponentCallbacks2;
 import android.content.Context;
 import android.content.res.Configuration;
@@ -67,7 +66,6 @@ import org.xmlrpc.android.ApiHelper;
 
 import java.io.File;
 import java.io.IOException;
-import java.lang.ref.WeakReference;
 import java.lang.reflect.Type;
 import java.security.GeneralSecurityException;
 import java.util.Date;
@@ -478,10 +476,6 @@ public class WordPress extends Application {
         return (getCurrentBlog() != null ? getCurrentBlog().getLocalTableBlogId() : -1);
     }
 
-    public static void signOutWordPressComAsyncWithProgressBar(Context context) {
-        new SignOutWordPressComAsync(context).executeOnExecutor(AsyncTask.THREAD_POOL_EXECUTOR);
-    }
-
     /**
      * Sign out from wpcom account
      */
@@ -524,41 +518,6 @@ public class WordPress extends Application {
 
         // dangerously delete all content!
         wpDB.dangerouslyDeleteAllContent();
-    }
-
-    public static class SignOutWordPressComAsync extends AsyncTask<Void, Void, Void> {
-        ProgressDialog mProgressDialog;
-        WeakReference<Context> mWeakContext;
-
-        public SignOutWordPressComAsync(Context context) {
-            mWeakContext = new WeakReference<Context>(context);
-        }
-
-        @Override
-        protected void onPreExecute() {
-            super.onPreExecute();
-            Context context = mWeakContext.get();
-            if (context != null) {
-                mProgressDialog = ProgressDialog.show(context, null, context.getText(R.string.signing_out), false);
-            }
-        }
-
-        @Override
-        protected Void doInBackground(Void... params) {
-            Context context = mWeakContext.get();
-            if (context != null) {
-                WordPressComSignOut(context);
-            }
-            return null;
-        }
-
-        @Override
-        protected void onPostExecute(Void aVoid) {
-            super.onPostExecute(aVoid);
-            if (mProgressDialog != null) {
-                mProgressDialog.dismiss();
-            }
-        }
     }
 
     public static void removeWpComUserRelatedData(Context context) {
